+++ packages/react-table/CHANGELOG.md
@@ -243,6 +243,124 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 **Note:** Version bump only for package @patternfly/react-table
 
+# [5.2.0-prerelease.48](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.47...@patternfly/react-table@5.2.0-prerelease.48) (2024-01-17)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# 5.2.0-prerelease.47 (2024-01-16)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.46](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.45...@patternfly/react-table@5.2.0-prerelease.46) (2024-01-15)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.45](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.44...@patternfly/react-table@5.2.0-prerelease.45) (2024-01-11)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.44](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.43...@patternfly/react-table@5.2.0-prerelease.44) (2024-01-10)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.43](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.42...@patternfly/react-table@5.2.0-prerelease.43) (2024-01-09)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.42](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.41...@patternfly/react-table@5.2.0-prerelease.42) (2024-01-08)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# 5.2.0-prerelease.41 (2024-01-05)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.40](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.39...@patternfly/react-table@5.2.0-prerelease.40) (2024-01-04)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.39](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.38...@patternfly/react-table@5.2.0-prerelease.39) (2024-01-03)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.38](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.37...@patternfly/react-table@5.2.0-prerelease.38) (2024-01-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.37](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.36...@patternfly/react-table@5.2.0-prerelease.37) (2023-12-20)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.36](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.35...@patternfly/react-table@5.2.0-prerelease.36) (2023-12-14)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.35](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.34...@patternfly/react-table@5.2.0-prerelease.35) (2023-12-11)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.34](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.33...@patternfly/react-table@5.2.0-prerelease.34) (2023-12-07)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.32...@patternfly/react-table@5.2.0-prerelease.33) (2023-12-06)
+
+### Bug Fixes
+
+- **Table:** make tooltip prop consistent ([#9803](https://github.com/patternfly/patternfly-react/issues/9803)) ([eb9693f](https://github.com/patternfly/patternfly-react/commit/eb9693f321dd4265390d2ac18698d8adc892a3fd))
+
+# [5.2.0-prerelease.32](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.31...@patternfly/react-table@5.2.0-prerelease.32) (2023-11-21)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.31](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.30...@patternfly/react-table@5.2.0-prerelease.31) (2023-11-17)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.30](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.29...@patternfly/react-table@5.2.0-prerelease.30) (2023-11-16)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.29](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.28...@patternfly/react-table@5.2.0-prerelease.29) (2023-11-14)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.28](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.27...@patternfly/react-table@5.2.0-prerelease.28) (2023-11-14)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.27](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.26...@patternfly/react-table@5.2.0-prerelease.27) (2023-11-14)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.26](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.25...@patternfly/react-table@5.2.0-prerelease.26) (2023-11-13)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.25](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.24...@patternfly/react-table@5.2.0-prerelease.25) (2023-11-10)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.24](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.23...@patternfly/react-table@5.2.0-prerelease.24) (2023-11-03)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.23](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.22...@patternfly/react-table@5.2.0-prerelease.23) (2023-11-03)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.22](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.21...@patternfly/react-table@5.2.0-prerelease.22) (2023-11-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.21](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.20...@patternfly/react-table@5.2.0-prerelease.21) (2023-11-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
+# [5.2.0-prerelease.20](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.19...@patternfly/react-table@5.2.0-prerelease.20) (2023-11-02)
+
+**Note:** Version bump only for package @patternfly/react-table
+
 # [5.2.0-prerelease.19](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-table@5.2.0-prerelease.18...@patternfly/react-table@5.2.0-prerelease.19) (2023-10-30)
 
 **Note:** Version bump only for package @patternfly/react-table